+++ Example/__tests__/index.ios.js
@@ -1,4 +1,3 @@
-import 'react-native';
 import React from 'react';
 import Index from '../index.ios.js';
 
@@ -6,7 +5,5 @@ import Index from '../index.ios.js';
 import renderer from 'react-test-renderer';
 
 it('renders correctly', () => {
-  const tree = renderer.create(
-    <Index />
-  );
+  const tree = renderer.create(<Index />);
 });