+++ Example/index.android.js
@@ -1,2 +1,0 @@
-import kontaktIoExample from './src/main';
-kontaktIoExample('android');