+++ Example/src/Example.android.tsx
@@ -7,8 +7,13 @@ import {
   ScrollView,
   TouchableOpacity,
 } from 'react-native';
-
 import Kontakt from 'react-native-kontaktio';
+import type { ColorValue } from 'react-native';
+import type {
+  ConfigType,
+  RegionType,
+  IBeaconAndroid,
+} from 'react-native-kontaktio';
 
 const {
   connect,
@@ -34,18 +39,25 @@ const {
   monitoringSyncInterval,
 } = Kontakt;
 
-const region1 = {
+const region1: RegionType = {
   identifier: 'Test beacons 1',
   uuid: 'B0702880-A295-A8AB-F734-031A98A512D3',
   major: 1,
   // no minor provided: will detect all minors
 };
 
-const region2 = {
+const region2: RegionType = {
   identifier: 'Test beacons 2',
   uuid: 'B0702880-A295-A8AB-F734-031A98A512D3',
   major: 2,
   // no minor provided: will detect all minors
+};
+
+type State = {
+  scanning: boolean;
+  beacons: Array<IBeaconAndroid>;
+  eddystones: Array<IBeaconAndroid>;
+  statusText: string | null;
 };
 
 /**
@@ -66,8 +78,8 @@ const region2 = {
  * sorted by their RSSI which reflects a measure of distance from the beacons
  * to your mobile phone.
  */
-export default class IBeaconExample extends Component {
-  state = {
+export default class IBeaconExample extends Component<{}, State> {
+  state: State = {
     scanning: false,
     beacons: [],
     eddystones: [],
@@ -76,24 +88,23 @@ export default class IBeaconExample extends Component {
 
   componentDidMount() {
     // Initialization, configuration and adding of beacon regions
-    connect(
-      'MY_KONTAKTIO_API_KEY',
-      [IBEACON, EDDYSTONE],
-    )
-      .then(() => configure({
-        scanMode: scanMode.BALANCED,
-        scanPeriod: scanPeriod.create({
-          activePeriod: 6000,
-          passivePeriod: 20000,
-        }),
-        activityCheckConfiguration: activityCheckConfiguration.DEFAULT,
-        forceScanConfiguration: forceScanConfiguration.MINIMAL,
-        monitoringEnabled: monitoringEnabled.TRUE,
-        monitoringSyncInterval: monitoringSyncInterval.DEFAULT,
-      }))
+    const config: ConfigType = {
+      scanMode: scanMode.BALANCED,
+      scanPeriod: scanPeriod.create({
+        activePeriod: 6000,
+        passivePeriod: 20000,
+      }),
+      activityCheckConfiguration: activityCheckConfiguration.DEFAULT,
+      forceScanConfiguration: forceScanConfiguration.MINIMAL,
+      monitoringEnabled: monitoringEnabled.TRUE,
+      monitoringSyncInterval: monitoringSyncInterval.DEFAULT,
+    };
+
+    connect('MY_KONTAKTIO_API_KEY', [IBEACON, EDDYSTONE])
+      .then(() => configure(config))
       .then(() => setBeaconRegions([region1, region2]))
-      .then(() => setEddystoneNamespace())
-      .catch(error => console.log('error', error));
+      .then(() => setEddystoneNamespace(null))
+      .catch((error) => console.log('error', error));
 
     // Beacon listeners
     DeviceEventEmitter.addListener(
@@ -102,7 +113,7 @@ export default class IBeaconExample extends Component {
         console.log('beaconDidAppear', newBeacon, region);
 
         this.setState({
-          beacons: this.state.beacons.concat(newBeacon)
+          beacons: this.state.beacons.concat(newBeacon),
         });
       }
     );
@@ -112,11 +123,11 @@ export default class IBeaconExample extends Component {
         console.log('beaconDidDisappear', lostBeacon, region);
 
         const { beacons } = this.state;
-        const index = beacons.findIndex(beacon =>
+        const index = beacons.findIndex((beacon) =>
           this._isIdenticalBeacon(lostBeacon, beacon)
         );
         this.setState({
-          beacons: beacons.reduce((result, val, ind) => {
+          beacons: beacons.reduce<Array<IBeaconAndroid>>((result, val, ind) => {
             // don't add disappeared beacon to array
             if (ind === index) return result;
             // add all other beacons to array
@@ -124,52 +135,52 @@ export default class IBeaconExample extends Component {
               result.push(val);
               return result;
             }
-          }, [])
+          }, []),
         });
       }
     );
     DeviceEventEmitter.addListener(
       'beaconsDidUpdate',
-      ({ beacons: updatedBeacons, region }) => {
+      ({
+        beacons: updatedBeacons,
+        region,
+      }: {
+        beacons: Array<IBeaconAndroid>;
+        region: RegionType;
+      }) => {
         console.log('beaconsDidUpdate', updatedBeacons, region);
 
         const { beacons } = this.state;
-        updatedBeacons.forEach(updatedBeacon => {
-          const index = beacons.findIndex(beacon =>
+        updatedBeacons.forEach((updatedBeacon) => {
+          const index = beacons.findIndex((beacon) =>
             this._isIdenticalBeacon(updatedBeacon, beacon)
           );
           this.setState({
-            beacons: beacons.reduce((result, val, ind) => {
-              // replace current beacon values for updatedBeacon, keep current value for others
-              ind === index ? result.push(updatedBeacon) : result.push(val);
-              return result;
-            }, [])
-          })
+            beacons: beacons.reduce<Array<IBeaconAndroid>>(
+              (result, val, ind) => {
+                // replace current beacon values for updatedBeacon, keep current value for others
+                ind === index ? result.push(updatedBeacon) : result.push(val);
+                return result;
+              },
+              []
+            ),
+          });
         });
       }
     );
 
     // Region listeners
-    DeviceEventEmitter.addListener(
-      'regionDidEnter',
-      ({ region }) => {
-        console.log('regionDidEnter', region);
-      }
-    );
-    DeviceEventEmitter.addListener(
-      'regionDidExit',
-      ({ region }) => {
-        console.log('regionDidExit', region);
-      }
-    );
+    DeviceEventEmitter.addListener('regionDidEnter', ({ region }) => {
+      console.log('regionDidEnter', region);
+    });
+    DeviceEventEmitter.addListener('regionDidExit', ({ region }) => {
+      console.log('regionDidExit', region);
+    });
 
     // Beacon monitoring listener
-    DeviceEventEmitter.addListener(
-      'monitoringCycle',
-      ({ status }) => {
-        console.log('monitoringCycle', status);
-      }
-    );
+    DeviceEventEmitter.addListener('monitoringCycle', ({ status }) => {
+      console.log('monitoringCycle', status);
+    });
 
     /*
      * Eddystone
@@ -181,25 +192,18 @@ export default class IBeaconExample extends Component {
         console.log('eddystoneDidAppear', eddystone, namespace);
 
         this.setState({
-          eddystones: this.state.eddystones.concat(eddystone)
+          eddystones: this.state.eddystones.concat(eddystone),
         });
       }
     );
 
-    DeviceEventEmitter.addListener(
-      'namespaceDidEnter',
-      ({ status }) => {
-        console.log('namespaceDidEnter', status);
-      }
-    );
+    DeviceEventEmitter.addListener('namespaceDidEnter', ({ status }) => {
+      console.log('namespaceDidEnter', status);
+    });
 
-    DeviceEventEmitter.addListener(
-      'namespaceDidExit',
-      ({ status }) => {
-        console.log('namespaceDidExit', status);
-      }
-    );
-
+    DeviceEventEmitter.addListener('namespaceDidExit', ({ status }) => {
+      console.log('namespaceDidExit', status);
+    });
   }
 
   componentWillUnmount() {
@@ -212,71 +216,100 @@ export default class IBeaconExample extends Component {
     startScanning()
       .then(() => this.setState({ scanning: true, statusText: null }))
       .then(() => console.log('started scanning'))
-      .catch(error => console.log('[startScanning]', error));
+      .catch((error) => console.log('[startScanning]', error));
   };
   _stopScanning = () => {
     stopScanning()
-      .then(() => this.setState({ scanning: false, beacons: [], statusText: null }))
+      .then(() =>
+        this.setState({ scanning: false, beacons: [], statusText: null })
+      )
       .then(() => console.log('stopped scanning'))
-      .catch(error => console.log('[stopScanning]', error));
+      .catch((error) => console.log('[stopScanning]', error));
   };
   _restartScanning = () => {
     restartScanning()
-      .then(() => this.setState({ scanning: true, beacons: [], statusText: null }))
+      .then(() =>
+        this.setState({ scanning: true, beacons: [], statusText: null })
+      )
       .then(() => console.log('restarted scanning'))
-      .catch(error => console.log('[restartScanning]', error));
+      .catch((error) => console.log('[restartScanning]', error));
   };
   _isScanning = () => {
     isScanning()
-      .then(result => {
-        this.setState({ statusText: `Device is currently ${result ? '' : 'NOT '}scanning.` });
+      .then((result) => {
+        this.setState({
+          statusText: `Device is currently ${result ? '' : 'NOT '}scanning.`,
+        });
         console.log('Is device scanning?', result);
       })
-      .catch(error => console.log('[isScanning]', error));
+      .catch((error) => console.log('[isScanning]', error));
   };
   _isConnected = () => {
     isConnected()
-      .then(result => {
-        this.setState({ statusText: `Device is ${result ? '' : 'NOT '}ready to scan beacons.` });
+      .then((result) => {
+        this.setState({
+          statusText: `Device is ${result ? '' : 'NOT '}ready to scan beacons.`,
+        });
         console.log('Is device connected?', result);
       })
-      .catch(error => console.log('[isConnected]', error));
+      .catch((error) => console.log('[isConnected]', error));
   };
   _getBeaconRegions = () => {
     getBeaconRegions()
-      .then(regions => console.log('regions', regions))
-      .catch(error => console.log('[getBeaconRegions]', error));
+      .then((regions) => console.log('regions', regions))
+      .catch((error) => console.log('[getBeaconRegions]', error));
   };
 
   /**
    * Helper function used to identify equal beacons
    */
-  _isIdenticalBeacon = (b1, b2) => (
-    (b1.identifier === b2.identifier) &&
-    (b1.uuid === b2.uuid) &&
-    (b1.major === b2.major) &&
-    (b1.minor === b2.minor)
-  );
+  _isIdenticalBeacon = (b1: IBeaconAndroid, b2: IBeaconAndroid) =>
+    b1.uniqueId === b2.uniqueId &&
+    b1.uuid === b2.uuid &&
+    b1.major === b2.major &&
+    b1.minor === b2.minor;
 
   _renderBeacons = () => {
     const colors = ['#F7C376', '#EFF7B7', '#F4CDED', '#A2C8F9', '#AAF7AF'];
 
-    return this.state.beacons.sort((a, b) => a.accuracy - b.accuracy).map((beacon, ind) => (
-      <View key={ind} style={[styles.beacon, {backgroundColor: colors[beacon.minor - 1]}]}>
-        <Text style={{fontWeight: 'bold'}}>{beacon.uniqueId}</Text>
-        <Text>Major: {beacon.major}, Minor: {beacon.minor}</Text>
-        <Text>Distance: {beacon.accuracy}, Proximity: {beacon.proximity}</Text>
-        <Text>Battery Power: {beacon.batteryPower}, TxPower: {beacon.txPower}</Text>
-        <Text>FirmwareVersion: {beacon.firmwareVersion}, Address: {beacon.uniqueId}</Text>
-      </View>
-    ), this);
+    return this.state.beacons
+      .sort(
+        (a: IBeaconAndroid, b: IBeaconAndroid) =>
+          parseInt(a.accuracy) - parseInt(b.accuracy)
+      )
+      .map(
+        (beacon, ind) => (
+          <View
+            style={[
+              styles.beacon,
+              { backgroundColor: colors[beacon.minor - 1] },
+            ]}
+          >
+            <Text style={{ fontWeight: 'bold' }}>{beacon.uniqueId}</Text>
+            <Text>
+              Major: {beacon.major}, Minor: {beacon.minor}
+            </Text>
+            <Text>
+              Distance: {beacon.accuracy}, Proximity: {beacon.proximity}
+            </Text>
+            <Text>
+              Battery Power: {beacon.batteryPower}, TxPower: {beacon.txPower}
+            </Text>
+            <Text>
+              FirmwareVersion: {beacon.firmwareVersion}, Address:{' '}
+              {beacon.uniqueId}
+            </Text>
+          </View>
+        ),
+        this
+      );
   };
 
   _renderEmpty = () => {
     const { scanning, beacons } = this.state;
     let text;
-    if (!scanning) text = "Start scanning to listen for beacon signals!";
-    if (scanning && !beacons.length) text = "No beacons detected yet...";
+    if (!scanning) text = 'Start scanning to listen for beacon signals!';
+    if (scanning && !beacons.length) text = 'No beacons detected yet...';
     return (
       <View style={styles.textContainer}>
         <Text style={styles.text}>{text}</Text>
@@ -293,8 +326,15 @@ export default class IBeaconExample extends Component {
     ) : null;
   };
 
-  _renderButton = (text, onPress, backgroundColor) => (
-    <TouchableOpacity style={[styles.button, { backgroundColor }]} onPress={onPress}>
+  _renderButton = (
+    text: string,
+    onPress: () => void,
+    backgroundColor: ColorValue
+  ) => (
+    <TouchableOpacity
+      style={[styles.button, { backgroundColor }]}
+      onPress={onPress}
+    >
       <Text>{text}</Text>
     </TouchableOpacity>
   );
@@ -314,11 +354,17 @@ export default class IBeaconExample extends Component {
           {this._renderButton('Is connected?', this._isConnected, '#f2a2a2')}
         </View>
         <View style={styles.buttonContainer}>
-          {this._renderButton('Beacon regions (log)', this._getBeaconRegions, '#F4ED5A')}
+          {this._renderButton(
+            'Beacon regions (log)',
+            this._getBeaconRegions,
+            '#F4ED5A'
+          )}
         </View>
         {this._renderStatusText()}
         <ScrollView>
-          {scanning && beacons.length ? this._renderBeacons() : this._renderEmpty()}
+          {scanning && beacons.length
+            ? this._renderBeacons()
+            : this._renderEmpty()}
         </ScrollView>
       </View>
     );