+++ Example/src/Placeholder.tsx
@@ -1,9 +1,4 @@
-import React from 'react';
-import {
-  View,
-  Text,
-  StyleSheet,
-} from 'react-native';
+import { View, Text, StyleSheet } from 'react-native';
 
 const Placeholder = () => (
   <View style={styles.container}>