+++ Example/src/main.ios.ts
@@ -1,0 +1,7 @@
+import { AppRegistry } from 'react-native';
+
+import Example from './Example.ios';
+
+export default function native() {
+  AppRegistry.registerComponent('KontaktIoSimpleTest', () => Example);
+}