+++ CHANGELOG.md
@@ -6,6 +6,11 @@ _The format is based on [Keep a Changelog](http://keepachangelog.com/) and this 
 
 ## [unreleased]
 
+### Changed
+
+- Migrated library to TypeScript
+- Upgraded Example app to use React Native 0.65.2
+
 ## [v2.10.0] - 2021-11-26
 
 ### Changed
@@ -154,8 +159,8 @@ _The format is based on [Keep a Changelog](http://keepachangelog.com/) and this 
 
 - Several bug fixes
 
-
-[unreleased]: https://github.com/Artirigo/react-native-kontaktio/compare/v2.9.0...master
+[unreleased]: https://github.com/Artirigo/react-native-kontaktio/compare/v2.10.0...master
+[v2.10.0]: https://github.com/Artirigo/react-native-kontaktio/compare/v2.9.0...v2.10.0
 [v2.9.0]: https://github.com/Artirigo/react-native-kontaktio/compare/v2.8.0...v2.9.0
 [v2.8.0]: https://github.com/Artirigo/react-native-kontaktio/compare/v2.7.2...v2.8.0
 [v2.7.2]: https://github.com/Artirigo/react-native-kontaktio/compare/v2.7.1...v2.7.2